Quick sprite module split
