+++ arcade/sprite/base.py
@@ -1,13 +1,5 @@
-"""
-This module manages all of the code around Sprites.
-
-For information on Spatial Hash Maps, see:
-https://www.gamedev.net/articles/programming/general-and-gameplay-programming/spatial-hashing-r2697/
-"""
-
 import math
 from math import sin, cos, radians
-import dataclasses
 from typing import (
     Any,
     Tuple,
@@ -18,29 +10,18 @@ from typing import (
     TYPE_CHECKING,
 )
 from pathlib import Path
-import PIL.Image
 
 import arcade
 from arcade.geometry_generic import get_angle_degrees
 from arcade import (
     load_texture,
     Texture,
-    make_soft_circle_texture,
-    make_circle_texture,
 )
 from arcade.color import BLACK
-from arcade.resources import resolve_resource_path
 from arcade.types import Color, RGBA, Point, PointList, PathOrTexture
-from arcade.texture import SolidColorTexture
-from arcade import cache, hitbox
 
 if TYPE_CHECKING:  # handle import cycle caused by type hinting
     from arcade.sprite_list import SpriteList
-
-FACE_RIGHT = 1
-FACE_LEFT = 2
-FACE_UP = 3
-FACE_DOWN = 4
 
 
 class PyMunk:
@@ -1047,323 +1028,3 @@ class Sprite:
 
         # noinspection PyTypeChecker
         return check_for_collision_with_list(self, sprite_list)
-
-
-@dataclasses.dataclass
-class AnimationKeyframe:
-    """
-    Used in animated sprites.
-    """
-    tile_id: int
-    duration: int
-    texture: Texture
-
-
-class AnimatedTimeBasedSprite(Sprite):
-    """
-    Sprite for platformer games that supports animations. These can
-    be automatically created by the Tiled Map Editor.
-    """
-    def __init__(
-            self,
-            path_or_texture: PathOrTexture = None,
-            center_x: float = 0.0,
-            center_y: float = 0.0,
-            scale: float = 1.0,
-    ):
-        super().__init__(
-            path_or_texture,
-            scale=scale,
-            center_x=center_x,
-            center_y=center_y,
-        )
-        self.cur_frame_idx = 0
-        self.frames: List[AnimationKeyframe] = []
-        self.time_counter = 0.0
-
-    def update_animation(self, delta_time: float = 1 / 60) -> None:
-        """
-        Logic for selecting the proper texture to use.
-        """
-        self.time_counter += delta_time
-        while self.time_counter > self.frames[self.cur_frame_idx].duration / 1000.0:
-            self.time_counter -= self.frames[self.cur_frame_idx].duration / 1000.0
-            self.cur_frame_idx += 1
-            if self.cur_frame_idx >= len(self.frames):
-                self.cur_frame_idx = 0
-            # source = self.frames[self.cur_frame].texture.image.source
-            cur_frame = self.frames[self.cur_frame_idx]
-            # print(f"Advance to frame {self.cur_frame_idx}: {cur_frame.texture.name}")
-            self.texture = cur_frame.texture
-
-
-class AnimatedWalkingSprite(Sprite):
-    """
-    Deprecated Sprite for platformer games that supports walking animations.
-    Make sure to call update_animation after loading the animations so the
-    initial texture can be set. Or manually set it.
-
-    It is highly recommended you create your own version of this class rather than
-    try to use this pre-packaged one.
-
-    For an example, see this section of the platformer tutorial:
-    :ref:`platformer_part_twelve`.
-    """
-    def __init__(
-        self,
-        scale: float = 1.0,
-        center_x: float = 0.0,
-        center_y: float = 0.0,
-    ):
-        super().__init__(
-            None,
-            scale=scale,
-            center_x=center_x,
-            center_y=center_y,
-        )
-        self.state = FACE_RIGHT
-        self.stand_right_textures: List[Texture] = []
-        self.stand_left_textures: List[Texture] = []
-        self.walk_left_textures: List[Texture] = []
-        self.walk_right_textures: List[Texture] = []
-        self.walk_up_textures: List[Texture] = []
-        self.walk_down_textures: List[Texture] = []
-        self.cur_texture_index = 0
-        self.texture_change_distance = 20
-        self.last_texture_change_center_x: float = 0.0
-        self.last_texture_change_center_y: float = 0.0
-
-    def update_animation(self, delta_time: float = 1 / 60) -> None:
-        """
-        Logic for selecting the proper texture to use.
-        """
-        x1 = self.center_x
-        x2 = self.last_texture_change_center_x
-        y1 = self.center_y
-        y2 = self.last_texture_change_center_y
-        distance = math.sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2)
-        texture_list: List[Texture] = []
-
-        change_direction = False
-        if (
-            self.change_x > 0
-            and self.change_y == 0
-            and self.state != FACE_RIGHT
-            and len(self.walk_right_textures) > 0
-        ):
-            self.state = FACE_RIGHT
-            change_direction = True
-        elif (
-            self.change_x < 0
-            and self.change_y == 0
-            and self.state != FACE_LEFT
-            and len(self.walk_left_textures) > 0
-        ):
-            self.state = FACE_LEFT
-            change_direction = True
-        elif (
-            self.change_y < 0
-            and self.change_x == 0
-            and self.state != FACE_DOWN
-            and len(self.walk_down_textures) > 0
-        ):
-            self.state = FACE_DOWN
-            change_direction = True
-        elif (
-            self.change_y > 0
-            and self.change_x == 0
-            and self.state != FACE_UP
-            and len(self.walk_up_textures) > 0
-        ):
-            self.state = FACE_UP
-            change_direction = True
-
-        if self.change_x == 0 and self.change_y == 0:
-            if self.state == FACE_LEFT:
-                self.texture = self.stand_left_textures[0]
-            elif self.state == FACE_RIGHT:
-                self.texture = self.stand_right_textures[0]
-            elif self.state == FACE_UP:
-                self.texture = self.walk_up_textures[0]
-            elif self.state == FACE_DOWN:
-                self.texture = self.walk_down_textures[0]
-
-        elif change_direction or distance >= self.texture_change_distance:
-            self.last_texture_change_center_x = self.center_x
-            self.last_texture_change_center_y = self.center_y
-
-            if self.state == FACE_LEFT:
-                texture_list = self.walk_left_textures
-                if texture_list is None or len(texture_list) == 0:
-                    raise RuntimeError(
-                        "update_animation was called on a sprite that doesn't have a "
-                        "list of walk left textures."
-                    )
-            elif self.state == FACE_RIGHT:
-                texture_list = self.walk_right_textures
-                if texture_list is None or len(texture_list) == 0:
-                    raise RuntimeError(
-                        "update_animation was called on a sprite that doesn't have a list of "
-                        "walk right textures."
-                    )
-            elif self.state == FACE_UP:
-                texture_list = self.walk_up_textures
-                if texture_list is None or len(texture_list) == 0:
-                    raise RuntimeError(
-                        "update_animation was called on a sprite that doesn't have a list of "
-                        "walk up textures."
-                    )
-            elif self.state == FACE_DOWN:
-                texture_list = self.walk_down_textures
-                if texture_list is None or len(texture_list) == 0:
-                    raise RuntimeError(
-                        "update_animation was called on a sprite that doesn't have a list of walk down textures."
-                    )
-
-            self.cur_texture_index += 1
-            if self.cur_texture_index >= len(texture_list):
-                self.cur_texture_index = 0
-
-            self.texture = texture_list[self.cur_texture_index]
-
-        if self._texture is None:
-            print("Error, no texture set")
-        else:
-            self.width = self._texture.width * self.scale
-            self.height = self._texture.height * self.scale
-
-
-def load_animated_gif(resource_name) -> AnimatedTimeBasedSprite:
-    """
-    Attempt to load an animated GIF as an :class:`AnimatedTimeBasedSprite`.
-
-    Many older GIFs will load with incorrect transparency for every
-    frame but the first. Until the Pillow library handles the quirks of
-    the format better, loading animated GIFs will be pretty buggy. A
-    good workaround is loading GIFs in another program and exporting them
-    as PNGs, either as sprite sheets or a frame per file.
-    """
-
-    file_name = resolve_resource_path(resource_name)
-    # print(file_name)
-    image_object = PIL.Image.open(file_name)
-    if not image_object.is_animated:
-        raise TypeError(f"The file {resource_name} is not an animated gif.")
-
-    sprite = AnimatedTimeBasedSprite()
-    for frame in range(image_object.n_frames):
-        image_object.seek(frame)
-        frame_duration = image_object.info['duration']
-        image = image_object.convert("RGBA")
-        texture = Texture(image)
-        texture.origin = f"{resource_name}|{frame}"
-        sprite.textures.append(texture)
-        sprite.frames.append(AnimationKeyframe(0, frame_duration, texture))
-
-    sprite.texture = sprite.textures[0]
-    return sprite
-
-
-class SpriteSolidColor(Sprite):
-    """
-    A rectangular sprite of the given ``width``, ``height``, and ``color``.
-
-    The texture is automatically generated instead of loaded from a
-    file. Internally only a single global texture is used for this
-    sprite type, so concerns about memory usage non-existent regardless
-    of size or number of sprite variations.
-
-    :param int width: Width of the sprite in pixels
-    :param int height: Height of the sprite in pixels
-    :param Color color: The color of the sprite as an RGB or RGBA tuple
-    """
-    _default_image = PIL.Image.new("RGBA", (32, 32), (255, 255, 255, 255))
-
-    def __init__(
-        self,
-        width: int,
-        height: int,
-        *,
-        center_x: float = 0,
-        center_y: float = 0,
-        color: Color = (255, 255, 255, 255),
-    ):
-        """
-        Create a solid-color rectangular sprite.
-        """
-        super().__init__(
-            SolidColorTexture("sprite_solid_color", width, height, self._default_image),
-            center_x=center_x,
-            center_y=center_y,
-        )
-        self.color = arcade.get_four_byte_color(color)
-
-
-class SpriteCircle(Sprite):
-    """
-    A circle of the specified `radius <https://simple.wikipedia.org/wiki/Radius>`_.
-
-    The texture is automatically generated instead of loaded from a
-    file.
-
-    There may be a stutter the first time a combination of ``radius``,
-    ``color``, and ``soft`` is used due to texture generation. All
-    subsequent calls for the same combination will run faster because
-    they will re-use the texture generated earlier.
-
-    For a gradient fill instead of a solid color, set ``soft`` to
-    ``True``. The circle will fade from an opaque center to transparent
-    at the edges.
-
-    :param int radius: Radius of the circle in pixels
-    :param Color color: The Color of the sprite as an RGB or RGBA tuple
-    :param bool soft: If ``True``, the circle will fade from an opaque
-                      center to transparent edges.
-    """
-    def __init__(self, radius: int, color: Color, soft: bool = False):
-        radius = int(radius)
-        diameter = radius * 2
-        color_rgba = arcade.get_four_byte_color(color)
-
-        # We are only creating white textures. The actual color is
-        # is applied in the shader through the sprite's color attribute.
-        # determine the texture's cache name.
-        if soft:
-            cache_name = cache.crate_str_from_values("circle_texture_soft", diameter, 255, 255, 255, 255)
-        else:
-            cache_name = cache.crate_str_from_values("circle_texture", diameter, 255, 255, 255, 255)
-
-        # Get existing texture from cache if possible
-        texture = cache.texture_cache.get_with_config(cache_name, hitbox.algo_simple)
-        if not texture:
-            if soft:
-                texture = make_soft_circle_texture(
-                    diameter,
-                    color=(255, 255, 255, 255),
-                    name=cache_name,
-                    hit_box_algorithm=hitbox.algo_simple,
-                )
-            else:
-                texture = make_circle_texture(
-                    diameter,
-                    color=(255, 255, 255, 255),
-                    name=cache_name,
-                )
-            cache.texture_cache.put(texture)
-
-        # apply results to the new sprite
-        super().__init__(texture)
-        self.color = color_rgba
-        self._points = self.texture.hit_box_points
-
-
-def get_distance_between_sprites(sprite1: Sprite, sprite2: Sprite) -> float:
-    """
-    Returns the distance between the center of two given sprites
-
-    :param Sprite sprite1: Sprite one
-    :param Sprite sprite2: Sprite two
-    :return: Distance
-    :rtype: float
-    """
-    return arcade.get_distance(*sprite1._position, *sprite2._position)
