+++ arcade/sprite/enums.py
@@ -1,0 +1,4 @@
+FACE_RIGHT = 1
+FACE_LEFT = 2
+FACE_UP = 3
+FACE_DOWN = 4
